Add PARALLELIZATION!
So, the idea is as follows:

Instead of refining the list of items to search sequentially, we split
the items in chunks, as many as CPUs the machine has, and we send each
chunk to a different Future so it does the refinement.

To be honest, I didn't find much improvement between this way of doing
the filtering and the sequential way, which it only means that I'm doing
something wrong. But still, it is cool to have it.