Fix the movement of the current choice selected through the choices list
There was a problem when moving the highlighted selection through the
choices list when the number of choices was smaller than the height of
the window displaying them. The selection would go out of the list of
choices instead of cycling to the top/bottom as expected.

Now it is fixed. The Window now takes in account the current len of the
filtered choices in order to move the selected position.