Drop intermediate searchs before render
This improves rendering because the app doesn't waste time
on searchs that are not final. For example, typing "a" and then "b"
will probably trigger two searchs, but we only want to display the
latest, "ab", so we can render the query in the prompt and wait until
the search for "ab" is complete, dropping "a" in the middle.

The rendering of the layout now works well in fullscreen, inline and
inside NeoVim